Updated Due to Recommendations
This contract originally didn't follow the check and effects pattern, this was not an issue since the transfer of funds was performed through transfer() which limits the amount of gas and makes reentrancy not possible. 
However, since the community recommended that for clarity check and effects was followed the contract was changed.